requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

opencv-python
numpy
pandas
openpyxl
Pillow
matplotlib
scikit-learn
seaborn
tqdm
torch
torchvision